fix: run extension
